hello-app/requirements.txt
django
python-keycloak
pyjwt
